fix: LEAP-859: Fix data manager breaking on displaying specific data (#5692)
The commit replaces usage of `toStudlyCaps` across several files with a
new function `normalizeCellAlias` to standardize cell alias
normalization across the app. This change ensures a consistent approach
to handling cell alias normalization.

`CellViews` is used only to chose the right way do display the column
and for aliases there are just a couple of variants that are apllicable
in that case. So that changing any other string will not break anything.

### PR fulfills these requirements
- [ ] Tests for the changes have been added/updated
- [ ] Docs have been added/updated
- [x] Best efforts were made to ensure docs/code are concise and
coherent (checked for spelling/grammatical errors, commented out code,
debug logs etc.)
- [x] Self-reviewed and ran all changes on a local instance

### Describe the reason for change
`toStudlyCaps` includes
[code](https://github.com/dleitee/strman/blob/master/src/tostudlycaps.js#L30)
```JavaScript
value.replace(/[-_\s]+(.)?/g, function (match, chr) {
  return toUpperCase(chr);
});
``` 
which does not work with values like `"smthng_"` as `chr` should be a
string but we get `undefined`. To prevent this we need to remove unsafe
trailing special characters before using this function.

### This change affects (describe how if yes)
- [ ] Performance
- [ ] Security
- [ ] UX

### Does this PR introduce a breaking change?
- [ ] Yes, and covered entirely by feature flag(s)
- [ ] Yes, and covered partially by feature flag(s)
- [x] No
- [ ] Not sure (briefly explain the situation below)

### What level of testing was included in the change?
- [ ] e2e (codecept)
- [ ] integration (cypress)
- [ ] unit (jest)

### Which logical domain(s) does this change affect?
`CellViews`

---------

Co-authored-by: robot-ci-heartex <robot-ci-heartex@users.noreply.github.com>
Co-authored-by: Julio Sgarbi <julio.sgarbi@hotmail.com>
Co-authored-by: yyassi-heartex <104568407+yyassi-heartex@users.noreply.github.com>
Co-authored-by: Gondragos <Gondragos@users.noreply.github.com>